(Compute_Source_Coverage): Add missing reset of SCOs to be discharged
When the PC for a trace has multiple slocs, SCOs for each sloc must be
discharged separately. This change adds a missing reset of the variables
indicating which SCOs to discharge, whose absence caused SCOs for one
sloc being incorrectly discharged with information related to another
sloc for the same trace.

Fixes N118-001