gnatcov TOR: minor rework of the testcase predicate names for QM importers
The use of lone "test", as in "is_test", was misleading, different from
other uses of "test" in other pieces of the material.

Rename "is_test_case" as "is_tc", "is_test_set" as "is_tcset" and
"is_test" as "is_tc_or_set", so the tc related predicate naming scheme
is consistent with what we already have for requirement predicates.

In the course of N117-037