gnatcov TOR: remove dummy tc_set descriptions expecting an immediate table
They were introduced before the insertion of traceability links
to make sure the tables start from the left of the page. They are
pointless now (the traceability links perform the necessary flush)
and actually annoying as they expect the tc table
to follow immediatly (like "see the following table:"), but the
traceability links are inserted in between.

Fix presentation glitches observed in the course of N117-037