Annotate each LRM section with SCA applicabiliy
Annotated the LRM table of contents with an indication of
applicability to Source Coverage Analysis.  Possible entries
were "yes", "no", and "partial".  For "no" and "partial",
a comment was added explaining the rational. For "yes", there should be
at least one TOR testcase for that section.  Links to testcases
will be supplied by a separate tool.
MB23-001