gnatcov TOR/LRM: for non-applicable sections, always output comment
Instead of the set of tests found matching the section, if any.
This is relevant for tests flagged as applying to redundant sections,
which we can decide to claim applicable or not. This is also useful
for sections conditionally applicable, e.g.depending on the DO level.
We might have tests for a section relevant at level A, and claimed
not applicable at level B.

MB23-001