Add syntax highlighting to assembly listings
Followup on N127-051

This patch adds a Highlighting package that define primivites to create
a stream of tokens. It also adjust all the disassembly-related code to
use such primitives and disassemblers to format the appropriate tokens.

Finally, the dump-cfg command has been enhanced to take advantage of
this facility to colorate the assembly listings in its output,
highlighting condition branch instructions in a special way.