gnatcov testsuite: don't propagate --target if not passed at toplevel
Absence of --target at the toplevel testsuite driver today propagates down
to test instances as --target=<host-platform>, which propagates further down
as gnatcov run --target=<host-triplet>.

But the typical gnatcov run interface for native users is not to pass
--target at all, and absence of --target at the toplevel testsuite driver
should do that. This is what this change implements.

	* testsuite.py: In absence of a --target argument, don't propagate
	any --target at all to test instances.
	* SUITE/tutils.py (xrun): In absence of a --target argument, don't
	propagate any --target at all to gnatcov run.

Followup on N425-027