gnatcov test: Remove #decl anchor on deferred constant declaration
The testcase description explicitly only checks that "it is not
marked uncovered when elaborated", and the testcase code indeed
has only positive expectations on the deferred declaration itself.

Some compilers don't emit precise enough info to let us infer '+'
on that declaration, and removing the anchor is strong enough to
detect what the testcase aims at detecting (an unexpected '-' will
show up as a test error).

Qualification runs with only =report checks don't make this kind
of distinction between '+' and '.'.

For N910-002