Flag statements of units without code as not covered
There are two cases where we do not see any executable code for a SCO:
  * the compiler did not generate any code for the SCO
  * the enclosing unit was not part of any executable submitted to gnatcov

In the first case, the enclosing unit is part of the link, and we
expect to see *some* code for it, and if we do see such code then we
can confidently mark the SCO as No_Code and exlude it from the scope of
coverage analysis

In the second case, however, we don't know whether or not the compiler
generated any code for the SCO, and the situation is a likely indication
that the user omitted some test. We therefore must conservatively mark the
SCO as Not_Covered and report a violation.

For NA17-035, spun off from N610-006