Use project files to find source file locations, if possible
When generating reports with annotated sources, gnatcov needs to read
source files and thus to locate them precisely on the file system. In
the case of empty pure units (pragma Pure), compilation units may
produce no code at all and thus their absolute path may not make it to
build artifacts. As a consequence for such units, gnatcov will not be
able to locate the source file and thus to generate a complete coverage
report.

This change enhance source file location detection so that it rely on
the project manager for such units.

Followup on O302-015