gnatcov: forward -P and -X options to gnatemu
The general idea is to expose the scenario vars map, currently private to
the 'porject' unit, from switches.ads so the run driver can reconstruct a
set of -X options to pass to gnatemu.

The scenario var map materializes what -X switches gnatcov received so
switches.ads is a sensible place.

The container package instance and the map object itself are renamed to
have designations more appropriate for external references. The map isn't
called Scenario_Variables to prevent a conflict with a gnatcoll entity.

	* project.adb: Move Scv_Maps and Scv_Map to switches.ads.
	Rename along the way. Remove Add_Scenario_Var.
	* project.ads: Remove Add_Scenario_Var.
	* switches.ads: Expose the former Scv_Maps as Key_Element_Maps
	and the former Scv_Map as S_Variables.
	* gnatcov.adb: On -X, fill S_Variables directly.
	* rundrv.adb (Gnatemu_Driver_Control_For): Forward the -P and -X
	switches we have received.

Part of N630-009