disa_thumb/disa_arm: consider LDM* as branches when writing r15
On ARM, a lot of instructions can modify r15 (the PC register) and thus
work as a branch instruction. Among them, LDM/LDMIA/LDMFD is often used
in function epilogues to both restore registers and return from a
function.

Detect these branchy LDM instructions properly in the disassemblers.
We'll take care of the other instructions writing PC in another commit
since they seem to be very uncommon.

Followup on O804-024