Do not set outcome status on successors if dest kind unknown
Do not mark basic blocks downstream of a conditional branch edge
as known to be pre- or post-outcome unless the destination kind
is known for that edge: a Condition destination is known to be
pre-outcome, and a Decision destination is known to be post-outcome.
A destination of kind Unknown may be either.

Fixes O709-028