Set fallthrough destination on indirect jumps
The destination of the target edge of an indirect call is not known
at static analysis time, so we leave it to No_PC. The destination of the
fallthrough edge, on the other hand, is known and needs to be set so that
static analysis can properly propagate information through such calls.
Likewise for indirect jumps.

Fix for OA14-010

Change-Id: I1c1556628833b339b8342642cc7e618dfbb39819