P107-039 Using libopcode for arm disass
* libopcodes_bind/dis_wrapper.*
Wrappers for the functionalities of libopcode.

* libopcodes_bind/dis_opcodes.ads
Bindings from the C wrappers to Ada.

* disa_arm.ad*
Disassemble arm binary to text with dis-asm interface.

* disa_symbolize.ads | traces_elf.ad*
Add a primitive to symbolizer to query the name of a symbol at specified
address.

* disa_common.ad*
Implement a callback to perform symbol lookup for disassemblers using
libopcodes.
Implement a common function to disassemble code through libopcode.

Change-Id: I026d51deddebf502aefb005f7c805844f384a534