Introduce an option to filter SCOs based on the src. file they come from
Some Ada projects use separate subprograms as a mean to unit-test
internal entities (i.e. entities private to a unit body). As multiple
testcases bring different versions of these separate subprograms, and as
these are materialized as multiple different homonym symbols, this
currently triggers consolidation errors.

This change brings in a new option to disregard SCOs that target
specific source files. This way, these source files and thus the
corresponding object code artefacts will not contribute to the source
coverage (thus, will not trigger consolidation issues).

In case of inlining, object code that conflict can nevertheless be
required for source coverage analysis. For instance, assume
library-level unit A contains a subprogram B and multiple variants of a
separate subprogram C to tests B. A is an unit of interest but we want
to exclude A.C from the coverage analysis.

Because of inlining, machine code for B code can go inside the symbols
that materializes C. In this case, the multiple (and different) symbols
for C will be considered for coverage analysis and thus will conflict
during trace-based consolidation. In this case, checkpoint-based
consolidation provides a way out: each separate variant is analyzed in
its own checkpoint, and then checkpoint consolidation will make it
possible to produce a report without any consolidation issue.

Change-Id: I56ecf4d35b8632be80c0f504636e319553bf136a
TN: OA27-059