Elf_Disassemblers: fix handling for instruction set lookup cache miss
This change fixes a bug that could occure when disassembling machine
code for architecture with multiple instruction sets (such as ARM).

There is a facility to determine which disassembler to use for specific
instructions. For performance, this mechanism has a cache that speeds
the ISA lookup for usual cases. However, on cache miss, the current
lookup just gave up and returned "sorry, no information about this
instruction" whereas it should perform a full lookup instead in this
case. This is now fixed.

Change-Id: I89cef7a77fa238a89cafed45e7674727736c5b64
TN: P309-026