Traces_Files: enhance error messages in Open_Trace_File
In particular, print a specific error message when a second header is
expected but not found. The previous behavior yielded misleading error
messages later on.

As suggested in code review, this change also refactors parsing to make
sure we always have two headers, only the first one describing an
information section.

Change-Id: I8ac697d393bab325123f9c51788a424e91eed9d6
TN: M115-031