Refactor traces files handling
So far, there was multiple ways of reading trace entries. The
Traces_Files packages provides a "low-level" API (open trace file, read
trace entry, …) and "high-level" ones (Read_Trace_File overloads):
call-backs for entries, trace base to import entries to, etc.

There are multiple places in GNATcoverage that implement the high-level
logic on top of the low-level one. Some handle all the required stuff
(loadaddr special trace entries, for instance) an some don't. The result
is incomplete and not maintainable code.

This commit introduces a single generic procedure that will be the only
high-level implementation that relies on a private low-level API. All
the other "high-level" processing will be done on top of this generic.
The increase in maintainability will ease the incoming shared objects
handling work.

Additionally, this removes the assumption that all special trace entries
are loadaddr events. Now, errors are raised when the kind of a special
trace entry is unknown or unexpected.

In order to achieve this, this change performs a sub-refactoring so that
the Read_Trace_Entry primitive deals with raw the trace entry data
structures from Qemu_Traces instead of the "high-level" one from Traces.

Change-Id: I4fd0fc52710ea67846327bd84e72ac4fe470d259
TN: OA05-051