Rundrv: big spring refactoring
This change turns the descriptive Rundrv.Config database into
imperative-style driver lookup subprograms so that we can be more
flexible wrt. command-line creation. This will enable us to implement
more complex rules such as "don't emit this argument in verbose mode".

This change also make the environment changes first-class citizen so
that we'll be able to display what environment variables are set for the
trace producer.

As a bonus, this gets rid of the Rundrv.State package and thus also rid
of the global variables it used to hold. Such a state is now passed as
regular parameters, making data flow hopefuly sound.

Change-Id: I4cf88f9467a6b255ae1523c913fde7fc5f5eb0c5