Files_Table: distinguish source files and library files
The checkpoint machinery needs to have all library files in the files
table streamed. However, we also need to support stubbing
configurations, where different source files with the same name in
different projects are excluded from units of interest. As a
consequence, we must not stream to checkpoint files the entries in the
files table that are not considered for coverage reports.

This commit makes it possible to both stream all library files and to
restrict source files streaming to only units of interests.

Change-Id: I4c1a5ac5cae9f6edbdd890232bc40583b0a57534
TN: P121-052