gnatcov qualkit: sync item # between TOR and STR
Add item # columns to the STR tables describing the
qualification environment. Pick sequences of identifiers
starting with 'e' for Environment items and 's' for Suite
items, placed in distinct tables.

Adjust the Item # values in the Operational Conditions
of Use accordingly, and add a brief note explaining that
these identifiers link with those found in the STR.

PB04-040 (gnatcov qualkit: match STR and Operational environment)

Change-Id: Ib60c77c7cd084b00df9984ce53b45cd16e6a4b82