CFG_Dump: output a dot comment to display the command-line used
It happens that someone generates a CFG, keeps the resulting dot file
and then forgets about the command line used, making it hard to
re-generate the CFG later on. This commit enhances the dot generation to
include a top-level comment that will contain the command-line that was
used to generate the dot file.

Change-Id: I476472984538f99ab3a948e8dc517bda1bfba88c
TN: N115-039