gnatcov TOR: fix req description and tc titles on 552_Iteration items
Simplify requirement expression and fix testing strategy description,
imported from the one for For loop but not corresponding to what we are
really doing for iterators. Add consistency to the various TC titles,
remove bogus sectioning markup.

In the course of adjustments issued for general comments received
on PA26-030.

Import change issued for TEC.PB12-004  (qualkit 3661 x86-windows, TOR)
on the opendo branch.