gnatcov TOR: clarify restrictions on non-local handlers
Add an explicit mention that they apply to the code subject to coverage
analysis. Facilitate use of controlled objects, which expand handlers
and trigger warnings from No_Exception_Handlers but should be acceptable
because these handlers aren't subject to coverage analysis.

In the course of PB04-008

Import change issued for TEC.PB12-004  (qualkit 3661 x86-windows, TOR)
on the opendo branch.