gnatcov PLANS: improve list of Referenced Documents
Fix title of the GNAT user's guide and indicate the GNAT
release version we are referring to.

Also indicate the gnatcov release version in the ref to
the GNATcoverage user's guide.

Import change(s) issued for
TEC.PB07-002 (qualkit 3661 x86-windows, PLANS)

Address comment #1 of QAC1.1 activity tracked on PB21-036