gnatcov qualkit: simplify ref to gnatemu version in Environment
Just make it an extra line in the list of items, to be removed
for native qualkits. Match the item # produced in the STR.

No need for a whole paragraph conditioned
on "for cross configurations only", heavy and possibly confusing
in native material.

In the course of merging changes issued for
P701-038 (qualkit 1434 P55-elf-windows)