Testsuite: replace --target and --RTS with --config with project files
Now that gnatcov supports the --config option, passing --target and
--RTS is not required anymore when there is project support. Update the
testsuite to do that. This allows testing the new option and avoids
the need to generate an auto-config file when the testsuite config file
can be used.

Also tune project file generation to tell GNATemulator the name of the
emulated board.

Change-Id: Iae067e07bebf9060aa121588c1ad15b30670df20
TN: Q106-008