gnatcov suite: remove tinfo from "thistest"
thistest.tinfo was used as a cache for SUITE.control.target_info().

Stag methods were using it to perform symbol name transformations,
dragging dependency on the instantiation of "thistest" by SUITE.context.

This turns out problematic as soon as we try to run with -S instance
for qualification, as we then try to pickle dump and reload Stag objects
when building documents, and trying to reinstate thistest outside of a
regular testsuite run doesn't work (and should not be needed).

This change removes the .tinfo component from thistest (SUITE/context.py)
and replaces references by calls to SUITE.control.target_info, sometimes
cached on a per module basis to prevent repeated calls.

The target_info function was already accessed directly in some places,
so this improves consistency overall.

The change also removes a couple of useless imports in the same area
(eg. of LANGINFO in xnexpanders and tutils), spotted while examining who
is using what.

Part of Q123-017

Change-Id: I1918a8fe3035acf19cd373ad1b33fbb63949211f