remove examples/hello
Calls for maintenance as any example (e.g. comments in Makefile,
was referring to "qemu" in the text displayed, now very obsolete,
...).

In addition, too simple to be of interest coverage-wise, and we
have other examples in the very basic category so no added value
on this front.

Part of PC20-006 (review gnatcov examples)

Change-Id: I6aefaf3d96dee1b3615c01ea6d52c6b561b52cbe