Traces_Files: replace Bad_File_Format exception with Fatal_Error
Compared to a Fatal_Error, this exception could be used to do some error
recovery. However, no code did that, so get rid of this exception and
use Fatal_Error everywhere. This commit also adds context (trace file
name) to the resulting error messages, so that the cause of errors is
hopefully easier to understand.

Change-Id: I13798d5d8be586e43cdc831e7dc1fcc55d799e8e
TN: Q329-015