Makefile.examples: use .list files extensions consistently
Instead of .alis for list of alis and .routines for list of
routines per test, as we now also have a .list for all tests
when consolidating with checkpoints. Just a bit simpler to
grasp for the user and to handle for the clean: target.

Part of PC20-006 (review gnatcov examples)

Change-Id: Ic6f0fd86158c10a3d22b6aa64613653e7f0e360b