gnatcov examples: rename UNIT_TESTS as AUNIT_TESTS
The TESTS variable is already expected to contain a list
of unit tests, just handwritten. The former UNIT_TESTS variable
was indeed intended to received a list of AUNIT based tests,
and in particular to come with a "harness" project file. The
name update just better reflects the current situation.

Part of PC20-006 (review gnatcov examples)

Change-Id: I978f15e450c41ef15fffde0d4444ae786c283b81