Add a wrapper type around mapped regions
This is a preliminary refactoring to prepare the incoming work on
supporting compressed ELF sections. As sections contents will sometimes
no longer be memory mapped from a file, we need an wrapper type that can
store either a memory mapping or an allocated buffer.

No functional change intended.

Change-Id: Ib429bf31a58de38d84598a484434c16e4aae82ec
TN: Q905-005