(Analyze_Conditional_Branch): Ignore cond branch from landing pad
At the beginning of exception handlers, a landing pad is generated by
the compiler that dispatches the exception to an appropriate user-level
handler.  The conditional branch instructions for these landing pads do
not correspond to any source level condition, and must be ignored for
source coverage analysis purposes. They are identified by the fact that
at least one of their edges branches to a basic block that has a call
to __gnat_begin_handler.

Fixes Q117-027

Change-Id: Ie03e26c3528439f39ee8615a03cf08cd0c448dce