gnatcov: reinstate the "prepare" run target
gnatcov run --target=prepare was meant to let users prepare
for an execution that will just append trace entries to a trace
file, typically by running qemu separately. The preparation
consists in creating the trace file and storing a trace header
there.

This was removed based on the belief that this facility wasn't
used any more. It actually still is relied upon by a few scripts
involved in runtime certification activities, some currently
ongoing.

In the course of QA16-036

Change-Id: I719372cf9268d6964ff91d7e6a427acaeede1e09