altrun/trace32: Force reconnect when the probe is already connected
This can happen when the testsuite crashes, for instance with an rlimit.
The probe is still trying to talk with the previous Trace32 instance
that was killed by rlimit.

When we start the testsuite again, Trace32 will ask for reset
confirmation on the GUI. We cannot have that in the automated test run,
so the AUTOCONNECT option tells Trace32 to reset the probe without
asking for a user interaction.

Change-Id: Ife04a97f463efdd4ea0fc3b97bd85b40b1f7ed92