altrun/trace32: Use t32usbchecker to reset the probe
This can be needed when the testsuite crashes, for instance with an
rlimit. The probe is still trying to talk with the previous Trace32
instance that was killed by rlimit.

It should be possble to achive the same result with the
CONNECTIONMODE=AUTOCONNECT parameter of the configuration, however this
parameter doesn't work with the SCREEN=OFF mode.

Part of O721-022.

Change-Id: Id44d5dcf38ee906da34b4ffa9b9b7bc5e03c8a54