Remove SCO raw table
In the LAL based instrumenter, SCO generation is done in a single pass,
unlike the situation in the compiler where we do a first syntactic pass
and then a second filtering pass after the semantics pass has
back-annotated the raw SCO table. We can therefore directly populate the
"final" SCO table.

Change-Id: I59afce271817aeeb429ac3136ac111700031bd91