gnatcov run: Refine selection of an executable vs project and eargs
When there is no executable on the command line for gnatcov
run, use the first earg if executable, then check if an executable
can be inferred from project files.

Compared to the previous strategy, we only use the first earg if
it is executable, which allows passing only options through eargs
when project files provide the executable.

Part of RC14-011

Change-Id: I3456ec1b72748ad6640aebc646e640a826013c90