Add a support/src/last_chance_trace32-stm32f7 subdir
Since we need to pass --board for the trace32-stm32f7 configuration,
which triggers the addition of the corresponding last_chance source
subdir to allow overriding.

We don't really need to override the last chance handler at this
stage (specialized RTS involved), so just provide a README file.

Part of S109-040

Change-Id: I83c613201068481b1ddf1f47f5fb5962ac699884