Refactor Add_Auto_Dump_Buffers
The goal of this refactoring is to make it possible in the future to
reference from mains only the subset of buffer units relevant for the
compilation closure of the main (see the new Buffer_Units_For_Closure
function). This will avoid generating mains that will dump coverage
buffers that are always empty.

This also prepares the road to instrument mains that are not units of
interest.

Change-Id: I3dcf90028ddd8a7c663e12c4a3807583d88f4dfd
TN: Q529-014