Split the gnatcov_rts project in two
The new gnatcov_rts_full project is now usable directly for native
programs. As it is a library project, it should work out of the box with
complex trees of user library projects.

On the other hand, the gnatcov_rts project contains the minimal units
that instrumented programs need to build properly, leaving the
transmission of coverage buffers to user code. It should be more in line
with projects running on cross platforms as it has no dependency on the
runtime.

Change-Id: I4711860966bbc5bf95f49de77f6b10d8839175d8
TN: Q529-014