Fix use and implementation of SCOV_Helper.run_test
It was introduced to factorize the search of "exception raised"
indications in the test program's outputs for either binary or
source traces, but was shortcircuited.

Testing revealed the need to have it return the name of the
file containing the process's output instead of the process descriptor.

Take the opportunity to introduce dummy versions of the SCOV_Helper
methods that need specializations, with a brief documentation of what
they should do and return - still to be improved)

This fixes the recent regression in
Qualif/Appendix/Testsuite/Selftest/assert-failure.

Change-Id: Ifd44a168925654734cafe1bf3a0c3f5dabebed9f
TN: S321-037