gnatcov suite: fix --gprmode to provide SCOV units of interest
--gprmode is intended to let testers request the use of a project
file to convey units of interest. This was to replace the use of
--scos in the original model, and is now also needed to direct the
instrumentation.

The current implementation was just not adding anything to
the project file regarding units of interest for SCOV based tests.

This change addresses that by constructing a CovControl object
internally, class explicitly designed to allow this for tests
dedicated to the verification of the project files support.

Part of S321-037

Change-Id: Ife9b6f1ff8898c2724f018a7d62ca59336f82d77