Support for conditional expectations on test(suite) options
Add support for a new "%opts" key for CTL lines in expectations,
so we can express different expected results for trace-mode=src vs
trace-mode=bin. This was deemed legitimate, at least temporarily,
for analysis on generic units in the course of S417-010.

The existence of a dedicated CTL key will make it easy to spot
where we have such differences, which we must remain careful not
to spread gratuitously because they incur documentation complications.

Change-Id: Ia9d4df2c0cd5218e6e11fa76ace4923499fdd4c3
TN: S417-010