Improve warning for inconsistent coverage method during ckpt. loading
When loading a CU from a checkpoint, if there is a discrepancy regarding
SCO provenance, it is very likely that the fingerprint will not match.
However, it is more helpful in this case to warn that the coverage
method is not consistent rather than complaining about the generic
"cannot merge coverage information". Also improve the wording when the
fingerprint does not match.

Change-Id: I1e04eb01230f48b714c812c6f7d90f6f24e9afe4
TN: S506-014