testsuite: Pass BUILDER.COMMON_CARGS through gprbuild -cargs
Instead of adding the options for all languages at the project
file generation level.

This way, tests that use custom projects don't have to fiddle
with arguments in various possible places to provide the options
(which was pretty cumbersome).

And in retrospect, it is very natural to have a variable named
<whatever>_CARGS operate through -cargs.

Tests that need fine control over the options used can request
not to pass the common ones by setting BUILDER.COMMON_CARGS to
None or to an empty iterable.

Improvement over the previous approach for S506-016

Change-Id: I05154d6e0ffddf9088ce79b5ff24d695a951ba2d