testsuite: Introduce GPRswitches
Introduce a GPRswitches class to represent the set of
project file related options passed to gnatcov commands
to designate units of interest.

The class has members mirroring the current xcov_instrument
interface for such options, and features an "as_strings" property
to translate the items as a list of command line switch strings.

Rework xcov_instrument to accept an instance of the class
instead of the individual arguments.

Adjust the CovControl class to expect an optional instance
instead of an optional list of switch strings.

Adjust SCOV_Helper to the CovControl changes.

Part of S523-037

Change-Id: I32a8e8c9a58a33d6bc5ef553f58f624d127a2bd6