SCOV_Helper: Straighten .scoptions definition and use
Rename mode_scoptions a mode_coverage_sco_options to stress the
point for which these options are meant to be involved. This clarifies
then intent and how it is legitimate for the bin and src versions to
differ so much.

The options just happen to also be used at gnatcov run time in the bin
traces mode, visible from the bin specific version of mode_execute.

Remove the self.scoptions attribute, too ambiguous to make sense.

Minor rework of gen_one_xcov_report() to use mode_coverage_sco_options()
directly, where it is clearly legitimate as we are running gnatcov
coverage exactly there.

Part of S523-037

Change-Id: I9c18d3f97397122011bbd46a00ea939a08c586e5