Do not generate statement SCOs for Pre/Post-Condition pragmas
In gnatcov:

These pragmas do not generate code at their point of appearance in
the flow of statements, so they are not relevant for statement coverage
(reinstating previously reverted change).

-----------

testsuite+Qualif: Handle pre/post not statement any more

Add testsuite support for "s/e" subsitutions, allowing anchors
to request "match statement uncovered expectations as expression
never evaluated".

This is used to refine expectations on expressions matched
with general anchors such as "eval" used in various kinds of
contexts. For instance:

	if A and then B then -- # eval
        ...
	pragma Precondition (A and then B); -- # eval :s/e:

Matched with expectations such as:

        --  /eval/ l- ## s-

in drivers not exercising the expression at all. The common
expectation assumes that the expression is enclosed in a
statement, which is generally the case. The refinement allows
using this common expectation, just hinting that there is no
enclosing statement in a particular context, and so that an
expression level diagnostic is expected instead of a statement
level one.

-----------

Qualif: Remove a couple of DC tests for !decision in precond context

Decision coverage tests for expressions that aren't decisions are
already of secondary importance. Those exercising such expressions
in pre/post condition context become entirely pointless when pre/post
conditions aren't even considered as statements any more.

This removes
 Qualif/Ada/decision/1_Core/Topologies/Flip/Pragmas/PreBody
 Qualif/Ada/decision/1_Core/Topologies/Flip/Pragmas/PreDecl

-----------

All for S606-037

Change-Id: I8e622e07fb95456cdfb254fc3bf5433fb14042e0