Fix --auto-dump-buffers with top-level exception handlers
The instrumentation for --auto-dump-buffers of procedures with top-level
exception handlers but no declaration of local variables was incorrect.
The call to dump the buffers happened in the exception handler, so no
source trace was produced in the no-exception case.

This commit fixes this, generating a handled statements list wrapper
whether there are declarations or not.

Change-Id: If9e27e62ec06b3b564e62c1c99787271ce7d616e
TN: S607-033