SC_Obligations.Load_SCO: warn when loading SCOs for the same unit twice
The only mechanism we had so far to prevent loading the same SCOs twice
was to check that we did not load the same ALI file name. This is not
enough as instrumentation can create SCOs for a source without an ALI
file to be involved, and an upcoming change will load
instrumentation-based SCOs before loading compiler-based ones.

To properly detect this situation, this patch introduces an extra check
that the main source for the ALI file loaded must not have a compilation
unit record yet. For this check to be complete, this commit also
registers the simple name of instrumented sources, so that the main
source can be found looking up its simple name.

Change-Id: I1a4d2c1813a76505ffb8783d3122b8f338c32907
Depends-On: I83990f8a7a8bc0eea7a3339741fa66a2658d30b8
TN: S529-038