Report decision violations for decisions in ignored statement SCOs
A pragma that generates freestanding decisions (such as
Pre/Postcondition) is not treated as a statement because it produces no
code at its place of occurrence in sources. The corresponding statement
SCOs are ignored for statement coverage purposes, so if the decision is
never evaluated, we can't depend on a statement violation to have been
emitted already, and we must report a violation at decision (or MC/DC)
level.

Followup on S606-037

Change-Id: I81ac668a298312729462ac98b450a284a2617484