testsuite: include SCOV_CARGS only for supported languages
to prevent compilation failures on tests which happen to involve
sources for which these options are not supported, such as tests
verifying absence of interference from unsupported languages in
mixed languages applications.

In addition to refining the computation performed by
gprbuild_cargs_with, nullify LANGINFO['C++'].sco_file to reflect
that we don't support the production of SCOs for C++.

This is a better representation of the reality and influences
control.KNOWN_LANGUAGES, which we can then better leverage in
the computation.

Fix for S625-008

Change-Id: Id06cb41507ee936ee8b8c7d7f3bbdfd2fe564049