cutils.py: arrange for lines_of(filename) to strip newlines
lines_of(filename) returns a list of all the lines in the
file. Currently, each line in the list has a newline terminating
character. This is more a nuisance than anything else, and callers
typically strip this newline or don't care (happen to work with
it and would still work without it).

As a new use which needs stripping is about to be introduced, this
change arranges to strip before returning, which better corresponds
to the original intent and facilitates use.

Part of work for Q529-014

Change-Id: I651281dd8a7a28a0d52a8b2f239b16742c2d1151