Correct instrumentation of SELECT statements
Implement special circuitry to instrument the alternatives
in a SELECT statements: the first statement in such construct
has special semantics, and we cannot insert a witness call before it.
Also, when instrumenting an ACCEPT statement in an accept alternative,
insert the witness at first position in the enclosed sequence of
statements.

For S625-020

Change-Id: If19aa839adad4e33591df1cc0a3325162e352b6f