Improve instrumentation of delay or entry call alternatives
When executing a SELECT statement, the timeout duration of any
delay alternative, as well as the evaluation of the actual parameters
for any entry call alternative, occur unconditionally before any
alternative is selected. So, the delay or entry call statements at
the beginning of such alternative must always be recorded as covered
as soon as execution of the SELECT statement begins. (This is assumed
to be consistent with the behaviour seen with binary traces).

To this effect, insert the witness calls for the relevant statements
in the parent sequence, immediately before the SELECT statement.

For S625-020

Change-Id: Iad5c8a331ad383cbeacc066831145909dec4813e