Instrument: fix sloc computation for pragma (P) decisions
The specification for low-level SCO tables (GNAT's scos.ads) mandates
that we associate the sloc of the "pragma" keyword to outer decisions
hosted by pragma statements. That was the intend of the implementation,
but a small bug (wrong traversal of syntactic parents to get the pragma
syntax node) yielded the wrong sloc. This commit fixes this computation.

Change-Id: Id7d3add9537ac47e2b6501f33c74a132a4d676b5
TN: S819-052