SC_Obligations: rework the SCO_Descriptor.Operands member
For decisions, having the full Operands member (which contains two SCO
references) does not make sense. This moves this member to the
Operator-only case and creates an Expression member (which is only one
SCO reference) for Decisions.

As a nice side effect, this reduces the level of nesting in
SCO_Descriptor's definition.

Change-Id: I7837f2d77de2b5982bd1ae41eca3ccbd9a193b1e
TN: S820-013