Instrument: always process subunits and stubs together
GNATcoverage's concept of a unit of interest maps to Ada's library
units. As a consequence, when a library unit that contains stubs is of
interest, GNATcoverage must consider that the related subunits are of
interest as well.

This commit enhances the instrumenter so that we instrument subunits as
soon as we see a stub and conversely.

Change-Id: Ie743b80d10a459e79270442d77434e124691c7c7
Depends-On: Idd143c7fcd6f13518cf26aa5b223e12e84b7d6fb
TN: S916-039