Instrument: do not consider AND/OR operators as logical operators
In binary trace mode, we do not consider AND/OR (non short circuit
operators) as logical, i.e. operators in decisions unless the
Short_Circuit_And_Or pragma is in effect.

The instrumenter cannot detect this pragma yet, not is is able to detect
the type of AND/OR operands. Instrumenting them as decisions when they
process something else than booleans generates invalid code, so for now,
conservalitely assume they are never short-circuiting.

Change-Id: I0029bf6e52208af8bf60036cfe9eb09dd708c60b
TN: S923-012