Project: register subunits in Unit_Map
Even though in principle we consider that subunits are not standalone
units (for instance when we check that all units of interest have LI
files), the testsuite still adds them to the list of units of interest.
We could decide in the future that it should not, but in the meantime,
pretend this should work.

Change-Id: Ibcb004305e88f0ad85927be270b347a920cde2ed
TN: S918-005