Instrument mains only after all other units
In order to generate code to emit coverage buffers, the instrumentation
of mains needs to know the list of all instrumented units. However,
since subunits are not independent units of interest, we generally
discover subunits during the instrumentation process.

Mains can only be subprogram units (they cannot be packages), so if we
instrument *all* units but mains first, and only then instrument mains,
we are sure that mains instrumentation will be able to get the
definitive list of coverage buffers.

Change-Id: Ic2e1d1bc7064b8425406c2ddc680da552080004b
TN: S918-005