testsuite: refine computation of units of interest
So we don't include subunits in lists of units
passed through project oriented mechanisms.

When computing "units of interest" from "sources of interest",
detect subunits as the sources with a '-' in the filename and
a text line starting with "separate", case insensitive.

To locate the actual source file corresponding to each
"source of interest" referenced from the expectations, arrange
for the expected-notes expander to expose a dictionary mapping
source-name to absolute paths, at a spot where we have everything
needed at hand.

SB04-016

Change-Id: I476aff254a331c669c2e34bef995ce45fa230f25